MP-234 :truck: Rename: domain.model을 domain.entity로 변경
