+++ src/main/java/kr/modusplant/domains/member/adapter/in/mapper/supers/MemberMapper.java
@@ -2,7 +2,7 @@ package kr.modusplant.domains.member.adapter.in.mapper.supers;
 
 import kr.modusplant.domains.member.adapter.in.request.MemberRegisterRequest;
 import kr.modusplant.domains.member.adapter.in.response.MemberResponse;
-import kr.modusplant.domains.member.domain.model.Member;
+import kr.modusplant.domains.member.domain.entity.Member;
 import kr.modusplant.domains.member.domain.vo.Nickname;
 
 public interface MemberMapper {
