+++ src/test/java/kr/modusplant/domains/member/test/utils/domain/MemberTestUtils.java
@@ -1,6 +1,6 @@
 package kr.modusplant.domains.member.test.utils.domain;
 
-import kr.modusplant.domains.member.domain.model.Member;
+import kr.modusplant.domains.member.domain.entity.Member;
 
 public interface MemberTestUtils extends MemberIdTestUtils, MemberStatusTestUtils, NicknameTestUtils, BirthDateTestUtils {
     default Member createMember() {
